requirements.txt
fastapi
uvicorn
streamlit
torch
transformers
llama-cpp-python
numpy
scikit-learn
pymupdf